Cleanup app-details and app-versions and merged them
